move HTMLCollectionOfBuilder to its own file
